Move codeblocks to new component and fix build issues.
