+++ src/components/index.ts
@@ -1,7 +1,7 @@
 export { default as ApiVersion } from "./ApiVersion.astro";
 export { default as Accordion } from "./Accordion.astro";
 export { default as Callout } from "./Callout.astro";
-export { default as Code } from "./Code.astro";
+export { default as CodeBlock } from "./CodeBlock.astro";
 export { default as FigmaEmbed } from "./FigmaEmbed/FigmaEmbed.astro";
 export { default as MinorVersion } from "./MinorVersion.astro";
 export { default as SdkVersion } from "./SdkVersion.astro";
