+++ src/content/docs/api/rs-api/reports-ja.mdx
@@ -47,7 +47,7 @@ GET https://dash.adjust.com/control-center/reports-service/report
 | `[dimension]__not_starts_with` | `String`  | 指定されたプレフィックスで始まらないディメンションに一致するサブストリング検索（大文字と小文字の区別なし）。                                                                                                                                              | `campaign__not_starts_with=ios`                                                                                                                                                                                                                                                                                                                                |
 | `[dimension]__ends_with`       | `String`  | 指定されたサフィックスで終わるディメンションに一致するサブストリング検索（大文字と小文字の区別なし）。                                                                                                                                                    | `campaign__ends_with=ios`                                                                                                                                                                                                                                                                                                                                      |
 | `[dimension]__not_ends_with`   | `String`  | 指定されたサフィックスで終わらないディメンションに一致するサブストリング検索（大文字と小文字の区別なし）。                                                                                                                                                | `campaign__not_ends_with=ios`                                                                                                                                                                                                                                                                                                                                  |
-| `[metric]__lt`                 | `String`  | 「この数値より小さい（<）」の条件で、任意の指標の値をフィルター。                                                                                                                                                                                         | `skad_installs__lt=100`                                                                                                                                                                                                                                                                                                                                        |
+| `[metric]__lt`                 | `String`  | 「この数値より小さい（\<）」の条件で、任意の指標の値をフィルター。                                                                                                                                                                                        | `skad_installs__lt=100`                                                                                                                                                                                                                                                                                                                                        |
 | `[metric]__lte`                | `String`  | 「この数値以下（≦）」の条件で任意の指標の値をフィルター。                                                                                                                                                                                                 | `skad_installs__lte=100`                                                                                                                                                                                                                                                                                                                                       |
 | `[metric]__gt`                 | `String`  | 「この数値より大きい（>）」の条件で任意の指標の値をフィルター。                                                                                                                                                                                           | `skad_installs__gt=100`                                                                                                                                                                                                                                                                                                                                        |
 | `[metric]__gte`                | `String`  | 「この数値以上（≧）」の条件で任意の指標の値をフィルター。                                                                                                                                                                                                 | `skad_installs__gte=100`                                                                                                                                                                                                                                                                                                                                       |
