+++ src/content/docs/sdk/android/migration/v4-to-v5.mdx
@@ -24,7 +24,7 @@ To start using SDK v5, you need to add it as a dependency in your project. To do
 
 1. Add the beta repository to your `build.gradle` file.
 
-   <Code title="build.gradle">
+   <CodeBlock title="build.gradle">
 
    ```groovy
    repositories {
@@ -34,12 +34,12 @@ To start using SDK v5, you need to add it as a dependency in your project. To do
    }
    ```
 
-   </Code>
+   </CodeBlock>
 
 2. Remove any older versions of the SDK you currently have in your `build.gradle` file.
 3. Add the following dependencies to your `build.gradle` file:
 
-   <Code title="build.gradle">
+   <CodeBlock title="build.gradle">
 
    ```groovy
    dependencies {
@@ -50,7 +50,7 @@ To start using SDK v5, you need to add it as a dependency in your project. To do
    }
    ```
 
-   </Code>
+   </CodeBlock>
 
 4. Run `mvn install` to install the dependencies.
 
@@ -72,14 +72,14 @@ In SDK v4, the initialization method is `Adjust.onCreate(AdjustConfig)`. This ha
 
 </MinorVersion>
 
-<Code>
+<CodeBlock>
 
 ```java
 - Adjust.onCreate(adjustConfig); // v4
 + Adjust.initSdk(adjustConfig); // v5
 ```
 
-</Code>
+</CodeBlock>
 
 ### [Configuration](configuration)
 
@@ -314,14 +314,14 @@ The method used for passing deep link data to the Adjust SDK has been renamed.
 
 Both methods accept the same arguments.
 
-<Code>
+<CodeBlock>
 
 ```java
 - Adjust.appWillOpenUrl(data, getApplicationContext()); //v4
 + Adjust.processDeeplink(data, getApplicationContext()); //v5
 ```
 
-</Code>
+</CodeBlock>
 
 #### [Deferred deep linking callback](deferred-callback)
 
@@ -403,14 +403,14 @@ In SDK v4, ad revenue sources are defined as constants in the `AdjustConfig` cla
 
 </MinorVersion>
 
-<Code>
+<CodeBlock>
 
 ```java
 - AdjustAdRevenue adjustAdRevenue = new AdjustAdRevenue(AdjustConfig.AD_REVENUE_APPLOVIN_MAX); // v4
 + AdjustAdRevenue adjustAdRevenue = new AdjustAdRevenue("applovin_max_sdk"); // v5
 ```
 
-</Code>
+</CodeBlock>
 
 <Table>
 
@@ -472,53 +472,53 @@ Check the table below to see how to configure your URL strategy in SDK v5.
 
 ##### [Examples](examples)
 
-<Code title="India URL strategy">
+<CodeBlock title="India URL strategy">
 
 ```java
 config.setUrlStrategy(Arrays.asList("adjust.net.in", "adjust.com"), true, false);
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="China URL strategy">
+<CodeBlock title="China URL strategy">
 
 ```java
 adjustConfig.setUrlStrategy(Arrays.asList("adjust.world", "adjust.com"), true, false);
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="China only URL strategy">
+<CodeBlock title="China only URL strategy">
 
 ```java
 adjustConfig.setUrlStrategy(Arrays.asList("adjust.cn"), true, false);
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="EU data residency">
+<CodeBlock title="EU data residency">
 
 ```java
 adjustConfig.setUrlStrategy(Arrays.asList("eu.adjust.com"), true, true);
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="Turkey data residency">
+<CodeBlock title="Turkey data residency">
 
 ```java
 adjustConfig.setUrlStrategy(Arrays.asList("tr.adjust.com"), true, true);
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="US data residency">
+<CodeBlock title="US data residency">
 
 ```java
 adjustConfig.setUrlStrategy(Arrays.asList("us.adjust.com"), true, true);
 ```
 
-</Code>
+</CodeBlock>
 
 #### [Disable third party sharing globally](disable-sharing-globally)
 
@@ -551,45 +551,45 @@ SDK v5 contains new getter methods for fetching install referrer information.
 
 </MinorVersion>
 
-<Code title="Google Play Store Referrer">
+<CodeBlock title="Google Play Store Referrer">
 
 ```java
 Adjust.getGooglePlayInstallReferrer(getApplicationContext(), onGooglePlayInstallReferrerReadListener);
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="Huawei Install Referrer">
+<CodeBlock title="Huawei Install Referrer">
 
 ```java
 AdjustHuaweiReferrer.getHuaweiAppGalleryInstallReferrer(getApplicationContext(), onHuaweiInstallReferrerReadListener);
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="Samsung Install Referrer">
+<CodeBlock title="Samsung Install Referrer">
 
 ```java
 AdjustSamsungReferrer.getSamsungInstallReferrer(getApplicationContext(), onSamsungInstallReferrerReadListener);
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="Vivo Install Referrer">
+<CodeBlock title="Vivo Install Referrer">
 
 ```java
 AdjustVivoReferrer.getVivoInstallReferrer(getApplicationContext(), onSamsungInstallReferrerReadListener);
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="Xiaomi Install Referrer">
+<CodeBlock title="Xiaomi Install Referrer">
 
 ```java
 AdjustXiaomiReferrer.getXiaomiInstallReferrer(getApplicationContext(), onSamsungInstallReferrerReadListener);
 ```
 
-</Code>
+</CodeBlock>
 
 #### [Device ID getters](device-id-getters)
 
@@ -599,53 +599,53 @@ In SDK v4, all device ID getter methods run synchronously. In SDK v5, these meth
 
 </MinorVersion>
 
-<Code title="ADID getter">
+<CodeBlock title="ADID getter">
 
 ```java
 Adjust.getAdId(onAdidReadListener);
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="Attribution information getter">
+<CodeBlock title="Attribution information getter">
 
 ```java
 Adjust.getAttribution(onAttributionReadListener);
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="OAID getter">
+<CodeBlock title="OAID getter">
 
 ```java
 AdjustOaid.getOaid(getApplicationContext(), onOaidReadListener);
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="Amazon ADID getter">
+<CodeBlock title="Amazon ADID getter">
 
 ```java
 Adjust.getAmazonAdId(getApplicationContext(), onAmazonAdIdReadListener);
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="SDK version getter">
+<CodeBlock title="SDK version getter">
 
 ```java
 Adjust.getSdkVersion(onSdkVersionReadListener);
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="Enabled status getter">
+<CodeBlock title="Enabled status getter">
 
 ```java
 Adjust.isEnabled(getApplicationContext(), onIsEnabledListener);
 ```
 
-</Code>
+</CodeBlock>
 
 ### [Plugins](plugins)
 
@@ -659,7 +659,7 @@ SDK v5 supports a Huawei Install Referrer plugin to record a device's Huawei Ins
 
 </MinorVersion>
 
-<Code title="build.gradle">
+<CodeBlock title="build.gradle">
 
 ```groovy
 dependencies {
@@ -667,7 +667,7 @@ dependencies {
 }
 ```
 
-</Code>
+</CodeBlock>
 
 #### [Meta Install Referrer plugin](mir-plugin)
 
@@ -679,7 +679,7 @@ SDK v5 supports a Meta Install Referrer plugin to record [Meta Install Referrer]
 
 1. Add the following dependency to your `build.gradle` file:
 
-   <Code title="build.gradle">
+   <CodeBlock title="build.gradle">
 
    ```groovy
    dependencies {
@@ -687,12 +687,12 @@ SDK v5 supports a Meta Install Referrer plugin to record [Meta Install Referrer]
    }
    ```
 
-   </Code>
+   </CodeBlock>
 
 2. Find your Meta app ID in your [App Dashboard](https://developers.facebook.com/apps). See Meta's [App Dashboard documentation](https://developers.facebook.com/docs/development/create-an-app/app-dashboard) for more information.
 3. Pass your App ID as a `String` argument to the `AdjustConfig.setFbAppId` method.
 
-   <Code highlight="{range: 2}">
+   <CodeBlock highlight="{range: 2}">
 
    ```java
    AdjustConfig config = new AdjustConfig(this, appToken, environment);
@@ -700,7 +700,7 @@ SDK v5 supports a Meta Install Referrer plugin to record [Meta Install Referrer]
    Adjust.initSdk(config);
    ```
 
-   </Code>
+   </CodeBlock>
 
 #### [Removed plugins](removed-plugins)
