+++ src/content/docs/sdk/ios/migration/v4-to-v5-ko.mdx
@@ -34,13 +34,13 @@ To start using SDK v5, you need to add it as a dependency in your XCode project.
 
 2. `Podfile`에 다음 줄을 추가합니다.
 
-   <Code translate="no" title="Podfile">
+   <CodeBlock title="Podfile">
 
    ```rb
    pod 'Adjust', :git => 'https://github.com/adjust/ios_sdk.git', :branch => 'v5'
    ```
 
-   </Code>
+   </CodeBlock>
 
 3. Run `pod install` to install the dependencies.
 
@@ -1114,17 +1114,18 @@ In SDK v4, ad revenue sources are defined as constants in the `ADJConfig` class.
 
 </MinorVersion>
 <Tabs>
-<Tab title="Swift" sync="swift"> <Code translate="no">
+<Tab title="Swift" sync="swift">
+<CodeBlock>
 
 ```swift
 - let adRevenue = ADJAdRevenue(source: ADJAdRevenueSourceAppLovinMAX); //v4
 + let adRevenue = ADJAdRevenue(source: "applovin_max_sdk") //v5
 ```
 
-</Code>
+</CodeBlock>
 </Tab>
 <Tab title="Objective-C" sync="objc">
-<Code translate="no">
+<CodeBlock>
 
 ```objc
 - ADJAdRevenue *revenue = [[ADJAdRevenue alloc]
@@ -1133,7 +1134,7 @@ In SDK v4, ad revenue sources are defined as constants in the `ADJConfig` class.
 +                               initWithSource:@"applovin_max_sdk"]; // v5
 ```
 
-</Code>
+</CodeBlock>
 </Tab>
 </Tabs>
 <Table>
@@ -1197,57 +1198,57 @@ Check the table below to see how to configure your URL strategy in SDK v5\.
 
 <Tabs>
 <Tab title="Swift" sync="swift">
-<Code translate="no" title="India URL strategy">
+<CodeBlock title="India URL strategy">
 
 ```swift
 let domain = ["adjust.net.in", "adjust.com"]
 adjustConfig.setUrlStrategy(domain, withSubdomains: true, andDataResidency: false)
 ```
 
-</Code>
-<Code translate="no" title="China URL strategy">
+</CodeBlock>
+<CodeBlock title="China URL strategy">
 
 ```swift
 let domain = ["adjust.world", "adjust.com"]
 adjustConfig.setUrlStrategy(domain, withSubdomains: true, andDataResidency: false)
 ```
 
-</Code>
-<Code translate="no" title="China only URL strategy">
+</CodeBlock>
+<CodeBlock title="China only URL strategy">
 
 ```swift
 let domain = ["adjust.cn"]
 adjustConfig.setUrlStrategy(domain, withSubdomains: true, andDataResidency: false)
 ```
 
-</Code>
-<Code translate="no" title="EU data residency">
+</CodeBlock>
+<CodeBlock title="EU data residency">
 
 ```swift
 let domain = ["eu.adjust.com"]
 adjustConfig.setUrlStrategy(domain, withSubdomains: true, andDataResidency: true)
 ```
 
-</Code>
-<Code translate="no" title="Turkey data residency">
+</CodeBlock>
+<CodeBlock title="Turkey data residency">
 
 ```swift
 let domain = ["tr.adjust.com"]
 adjustConfig.setUrlStrategy(domain, withSubdomains: true, andDataResidency: true)
 ```
 
-</Code>
-<Code translate="no" title="US data residency">
+</CodeBlock>
+<CodeBlock title="US data residency">
 
 ```swift
 let domain = ["us.adjust.com"]
 adjustConfig.setUrlStrategy(domain, withSubdomains: true, andDataResidency: true)
 ```
 
-</Code>
+</CodeBlock>
 </Tab>
 <Tab title="Objective-C" sync="objc">
-<Code translate="no" title="India URL strategy">
+<CodeBlock title="India URL strategy">
 
 ```objc
 NSArray *domain = [[NSArray alloc]
@@ -1255,8 +1256,8 @@ NSArray *domain = [[NSArray alloc]
 [adjustConfig setUrlStrategy:domain withSubdomains:YES andDataResidency:NO];
 ```
 
-</Code>
-<Code translate="no" title="China URL strategy">
+</CodeBlock>
+<CodeBlock title="China URL strategy">
 
 ```objc
 NSArray *domain = [[NSArray alloc]
@@ -1264,8 +1265,8 @@ NSArray *domain = [[NSArray alloc]
 [adjustConfig setUrlStrategy:domain withSubdomains:YES andDataResidency:NO];
 ```
 
-</Code>
-<Code translate="no" title="China only URL strategy">
+</CodeBlock>
+<CodeBlock title="China only URL strategy">
 
 ```objc
 NSArray *domain = [[NSArray alloc]
@@ -1273,8 +1274,8 @@ NSArray *domain = [[NSArray alloc]
 [adjustConfig setUrlStrategy:domain withSubdomains:YES andDataResidency:NO];
 ```
 
-</Code>
-<Code translate="no" title="EU data residency">
+</CodeBlock>
+<CodeBlock title="EU data residency">
 
 ```objc
 NSArray *domain = [[NSArray alloc]
@@ -1282,8 +1283,8 @@ NSArray *domain = [[NSArray alloc]
 [adjustConfig setUrlStrategy:domain withSubdomains:YES andDataResidency:YES];
 ```
 
-</Code>
-<Code translate="no" title="Turkey data residency">
+</CodeBlock>
+<CodeBlock title="Turkey data residency">
 
 ```objc
 NSArray *domain = [[NSArray alloc]
@@ -1291,8 +1292,8 @@ NSArray *domain = [[NSArray alloc]
 [adjustConfig setUrlStrategy:domain withSubdomains:YES andDataResidency:YES];
 ```
 
-</Code>
-<Code translate="no" title="US data residency">
+</CodeBlock>
+<CodeBlock title="US data residency">
 
 ```objc
 NSArray *domain = [[NSArray alloc]
@@ -1300,7 +1301,7 @@ NSArray *domain = [[NSArray alloc]
 [adjustConfig setUrlStrategy:domain withSubdomains:YES andDataResidency:YES];
 ```
 
-</Code>
+</CodeBlock>
 </Tab>
 </Tabs>
 
@@ -1677,7 +1678,8 @@ In SDK v4, all device information getter methods run synchronously. In SDK v5, t
 
 </MinorVersion>
 <Tabs>
-<Tab title="Swift" sync="swift"> <Code translate="no" title="ADID getter">
+<Tab title="Swift" sync="swift">
+<CodeBlock title="ADID getter">
 
 ```swift
 Adjust.adid() { adid in
@@ -1685,8 +1687,8 @@ Adjust.adid() { adid in
 };
 ```
 
-</Code>
-<Code translate="no" title="IDFA getter">
+</CodeBlock>
+<CodeBlock title="IDFA getter">
 
 ```swift
 Adjust.idfa() { idfa in
@@ -1694,8 +1696,8 @@ Adjust.idfa() { idfa in
 };
 ```
 
-</Code>
-<Code translate="no" title="IDFV getter">
+</CodeBlock>
+<CodeBlock title="IDFV getter">
 
 ```swift
 Adjust.idfv() { idfv in
@@ -1703,8 +1705,8 @@ Adjust.idfv() { idfv in
 };
 ```
 
-</Code>
-<Code translate="no" title="SDK version getter">
+</CodeBlock>
+<CodeBlock title="SDK version getter">
 
 ```swift
 Adjust.sdkVersion() { sdkVersion in
@@ -1712,8 +1714,8 @@ Adjust.sdkVersion() { sdkVersion in
 };
 ```
 
-</Code>
-<Code translate="no" title="Enabled status getter">
+</CodeBlock>
+<CodeBlock title="Enabled status getter">
 
 ```swift
 Adjust.isEnabled() { isEnabled in
@@ -1721,8 +1723,8 @@ Adjust.isEnabled() { isEnabled in
 };
 ```
 
-</Code>
-<Code translate="no" title="Attribution getter">
+</CodeBlock>
+<CodeBlock title="Attribution getter">
 
 ```swift
 Adjust.attribution() { attribution in
@@ -1730,10 +1732,10 @@ Adjust.attribution() { attribution in
 }
 ```
 
-</Code>
+</CodeBlock>
 </Tab>
 <Tab title="Objective-C" sync="objc">
-<Code translate="no" title="ADID getter">
+<CodeBlock title="ADID getter">
 
 ```objc
 [Adjust adidWithCompletionHandler:^(NSString * _Nullable adid) {
@@ -1741,8 +1743,8 @@ Adjust.attribution() { attribution in
 }];
 ```
 
-</Code>
-<Code translate="no" title="IDFA getter">
+</CodeBlock>
+<CodeBlock title="IDFA getter">
 
 ```objc
 [Adjust idfaWithCompletionHandler:^(NSString * _Nullable idfa) {
@@ -1750,8 +1752,8 @@ Adjust.attribution() { attribution in
 }];
 ```
 
-</Code>
-<Code translate="no" title="IDFV getter">
+</CodeBlock>
+<CodeBlock title="IDFV getter">
 
 ```objc
 [Adjust idfvWithCompletionHandler:^(NSString * _Nullable idfv) {
@@ -1759,8 +1761,8 @@ Adjust.attribution() { attribution in
 }];
 ```
 
-</Code>
-<Code translate="no" title="SDK version getter">
+</CodeBlock>
+<CodeBlock title="SDK version getter">
 
 ```objc
 [Adjust sdkVersionWithCompletionHandler:^(NSString * _Nullable sdkVersion) {
@@ -1768,8 +1770,8 @@ Adjust.attribution() { attribution in
 }];
 ```
 
-</Code>
-<Code translate="no" title="Enabled status getter">
+</CodeBlock>
+<CodeBlock title="Enabled status getter">
 
 ```objc
 [Adjust isEnabledWithCompletionHandler:^(BOOL isEnabled) {
@@ -1777,8 +1779,8 @@ Adjust.attribution() { attribution in
 }];
 ```
 
-</Code>
-<Code translate="no" title="Attribution getter">
+</CodeBlock>
+<CodeBlock title="Attribution getter">
 
 ```objc
 [Adjust attributionWithCompletionHandler:^(ADJAttribution * _Nullable attribution) {
@@ -1786,7 +1788,7 @@ Adjust.attribution() { attribution in
 }
 ```
 
-</Code>
+</CodeBlock>
 </Tab>
 </Tabs>
