+++ src/content/docs/sdk/ios/migration/v4-to-v5.mdx
@@ -33,13 +33,13 @@ To start using SDK v5, you need to add it as a dependency in your XCode project.
 1. Remove any older versions of the SDK you currently have in your `Podfile`.
 2. Add the following line to your `Podfile`:
 
-   <Code title="Podfile">
+   <CodeBlock title="Podfile">
 
    ```rb
    pod 'Adjust', :git => 'https://github.com/adjust/ios_sdk.git', :branch => 'v5'
    ```
 
-   </Code>
+   </CodeBlock>
 
 3. Run `pod install` to install the dependencies.
 
@@ -1137,19 +1137,19 @@ In SDK v4, ad revenue sources are defined as constants in the `ADJConfig` class.
 <Tabs>
 <Tab title="Swift" sync="swift">
 
-<Code>
+<CodeBlock>
 
 ```swift
 - let adRevenue = ADJAdRevenue(source: ADJAdRevenueSourceAppLovinMAX); //v4
 + let adRevenue = ADJAdRevenue(source: "applovin_max_sdk") //v5
 ```
 
-</Code>
+</CodeBlock>
 
 </Tab>
 <Tab title="Objective-C" sync="objc">
 
-<Code>
+<CodeBlock>
 
 ```objc
 - ADJAdRevenue *revenue = [[ADJAdRevenue alloc]
@@ -1158,7 +1158,7 @@ In SDK v4, ad revenue sources are defined as constants in the `ADJConfig` class.
 +                               initWithSource:@"applovin_max_sdk"]; // v5
 ```
 
-</Code>
+</CodeBlock>
 
 </Tab>
 </Tabs>
@@ -1225,64 +1225,64 @@ Check the table below to see how to configure your URL strategy in SDK v5.
 <Tabs>
 <Tab title="Swift" sync="swift">
 
-<Code title="India URL strategy">
+<CodeBlock title="India URL strategy">
 
 ```swift
 let domain = ["adjust.net.in", "adjust.com"]
 adjustConfig.setUrlStrategy(domain, withSubdomains: true, andDataResidency: false)
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="China URL strategy">
+<CodeBlock title="China URL strategy">
 
 ```swift
 let domain = ["adjust.world", "adjust.com"]
 adjustConfig.setUrlStrategy(domain, withSubdomains: true, andDataResidency: false)
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="China only URL strategy">
+<CodeBlock title="China only URL strategy">
 
 ```swift
 let domain = ["adjust.cn"]
 adjustConfig.setUrlStrategy(domain, withSubdomains: true, andDataResidency: false)
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="EU data residency">
+<CodeBlock title="EU data residency">
 
 ```swift
 let domain = ["eu.adjust.com"]
 adjustConfig.setUrlStrategy(domain, withSubdomains: true, andDataResidency: true)
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="Turkey data residency">
+<CodeBlock title="Turkey data residency">
 
 ```swift
 let domain = ["tr.adjust.com"]
 adjustConfig.setUrlStrategy(domain, withSubdomains: true, andDataResidency: true)
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="US data residency">
+<CodeBlock title="US data residency">
 
 ```swift
 let domain = ["us.adjust.com"]
 adjustConfig.setUrlStrategy(domain, withSubdomains: true, andDataResidency: true)
 ```
 
-</Code>
+</CodeBlock>
 
 </Tab>
 <Tab title="Objective-C" sync="objc">
 
-<Code title="India URL strategy">
+<CodeBlock title="India URL strategy">
 
 ```objc
 NSArray *domain = [[NSArray alloc]
@@ -1290,9 +1290,9 @@ NSArray *domain = [[NSArray alloc]
 [adjustConfig setUrlStrategy:domain withSubdomains:YES andDataResidency:NO];
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="China URL strategy">
+<CodeBlock title="China URL strategy">
 
 ```objc
 NSArray *domain = [[NSArray alloc]
@@ -1300,9 +1300,9 @@ NSArray *domain = [[NSArray alloc]
 [adjustConfig setUrlStrategy:domain withSubdomains:YES andDataResidency:NO];
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="China only URL strategy">
+<CodeBlock title="China only URL strategy">
 
 ```objc
 NSArray *domain = [[NSArray alloc]
@@ -1310,9 +1310,9 @@ NSArray *domain = [[NSArray alloc]
 [adjustConfig setUrlStrategy:domain withSubdomains:YES andDataResidency:NO];
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="EU data residency">
+<CodeBlock title="EU data residency">
 
 ```objc
 NSArray *domain = [[NSArray alloc]
@@ -1320,9 +1320,9 @@ NSArray *domain = [[NSArray alloc]
 [adjustConfig setUrlStrategy:domain withSubdomains:YES andDataResidency:YES];
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="Turkey data residency">
+<CodeBlock title="Turkey data residency">
 
 ```objc
 NSArray *domain = [[NSArray alloc]
@@ -1330,9 +1330,9 @@ NSArray *domain = [[NSArray alloc]
 [adjustConfig setUrlStrategy:domain withSubdomains:YES andDataResidency:YES];
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="US data residency">
+<CodeBlock title="US data residency">
 
 ```objc
 NSArray *domain = [[NSArray alloc]
@@ -1340,7 +1340,7 @@ NSArray *domain = [[NSArray alloc]
 [adjustConfig setUrlStrategy:domain withSubdomains:YES andDataResidency:YES];
 ```
 
-</Code>
+</CodeBlock>
 
 </Tab>
 </Tabs>
@@ -1726,7 +1726,7 @@ In SDK v4, all device information getter methods run synchronously. In SDK v5, t
 <Tabs>
 <Tab title="Swift" sync="swift">
 
-<Code title="ADID getter">
+<CodeBlock title="ADID getter">
 
 ```swift
 Adjust.adid() { adid in
@@ -1734,9 +1734,9 @@ Adjust.adid() { adid in
 };
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="IDFA getter">
+<CodeBlock title="IDFA getter">
 
 ```swift
 Adjust.idfa() { idfa in
@@ -1744,9 +1744,9 @@ Adjust.idfa() { idfa in
 };
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="IDFV getter">
+<CodeBlock title="IDFV getter">
 
 ```swift
 Adjust.idfv() { idfv in
@@ -1754,9 +1754,9 @@ Adjust.idfv() { idfv in
 };
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="SDK version getter">
+<CodeBlock title="SDK version getter">
 
 ```swift
 Adjust.sdkVersion() { sdkVersion in
@@ -1764,9 +1764,9 @@ Adjust.sdkVersion() { sdkVersion in
 };
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="Enabled status getter">
+<CodeBlock title="Enabled status getter">
 
 ```swift
 Adjust.isEnabled() { isEnabled in
@@ -1774,9 +1774,9 @@ Adjust.isEnabled() { isEnabled in
 };
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="Attribution getter">
+<CodeBlock title="Attribution getter">
 
 ```swift
 Adjust.attribution() { attribution in
@@ -1784,12 +1784,12 @@ Adjust.attribution() { attribution in
 }
 ```
 
-</Code>
+</CodeBlock>
 
 </Tab>
 <Tab title="Objective-C" sync="objc">
 
-<Code title="ADID getter">
+<CodeBlock title="ADID getter">
 
 ```objc
 [Adjust adidWithCompletionHandler:^(NSString * _Nullable adid) {
@@ -1797,9 +1797,9 @@ Adjust.attribution() { attribution in
 }];
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="IDFA getter">
+<CodeBlock title="IDFA getter">
 
 ```objc
 [Adjust idfaWithCompletionHandler:^(NSString * _Nullable idfa) {
@@ -1807,9 +1807,9 @@ Adjust.attribution() { attribution in
 }];
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="IDFV getter">
+<CodeBlock title="IDFV getter">
 
 ```objc
 [Adjust idfvWithCompletionHandler:^(NSString * _Nullable idfv) {
@@ -1817,9 +1817,9 @@ Adjust.attribution() { attribution in
 }];
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="SDK version getter">
+<CodeBlock title="SDK version getter">
 
 ```objc
 [Adjust sdkVersionWithCompletionHandler:^(NSString * _Nullable sdkVersion) {
@@ -1827,9 +1827,9 @@ Adjust.attribution() { attribution in
 }];
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="Enabled status getter">
+<CodeBlock title="Enabled status getter">
 
 ```objc
 [Adjust isEnabledWithCompletionHandler:^(BOOL isEnabled) {
@@ -1837,9 +1837,9 @@ Adjust.attribution() { attribution in
 }];
 ```
 
-</Code>
+</CodeBlock>
 
-<Code title="Attribution getter">
+<CodeBlock title="Attribution getter">
 
 ```objc
 [Adjust attributionWithCompletionHandler:^(ADJAttribution * _Nullable attribution) {
@@ -1847,7 +1847,7 @@ Adjust.attribution() { attribution in
 }
 ```
 
-</Code>
+</CodeBlock>
 
 </Tab>
 </Tabs>
